+++ CHANGELOG.md
@@ -18,3 +18,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `core` module
 - Added `mc` module
 - Added `util` module
+- Added support for 8 recipe types
+  - `blasting`
+  - `crafting_shaped`
+  - `crafting_shapeless`
+  - `crafting_transmute`
+  - `smelting`
+  - `smithing_transform`
+  - `smithing_trim`
+  - `smoking`